Experimenting with framework to writing/reading "tar" files (my own format), also capable of doing zip
This commit adds abstractions on top of memory mapped files, file descriptors, pipes and compression functions.

It also adds a "tar" writer/reader but with a different format. The format consists of multiple files serialized in sequence, and a
tail block containing how many files were added, their names, sizes and offsets in the file.

It also adds "mycat" application which reads a file name from the argument and outputs it to stdout - in C++.

Lastly, it can also build a library called libdecompress which can be used with LD_PRELOAD before calling "mycat".
That library will intercept popen64 and try to find the file in a tar file (passed as an environment variable).
It will map that file into memory and serve that to the user of popen.